Fixing a bug where some versions wouldn't get notified for updates
Because of the change in site structure, but still need to notify the users of 0.0.6 that there is a new update, this is a simple solution that will tell if there is a new update in the console. In 0.1 and higher we are using a new endpoint on our api systems.